Move generator helpers to their language specific directories
