+++ crates/cxx-qt-gen/src/generator/cpp/mod.rs
@@ -19,10 +19,14 @@ pub mod threading;
 
 use std::collections::BTreeSet;
 
+use crate::generator::cpp::fragment::CppNamedType;
+use crate::naming::cpp::syn_type_to_cpp_type;
+use crate::naming::TypeNames;
 use crate::{generator::structuring, parser::Parser};
 use externcxxqt::GeneratedCppExternCxxQtBlocks;
 use qobject::GeneratedCppQObject;
-use syn::Result;
+use syn::spanned::Spanned;
+use syn::{Error, FnArg, ForeignItemFn, Pat, PatIdent, PatType, Result};
 
 /// Representation of the generated C++ code for a group of QObjects
 pub struct GeneratedCppBlocks {
@@ -138,4 +142,37 @@ mod tests {
         let cpp = GeneratedCppBlocks::from(&parser).unwrap();
         assert_eq!(cpp.qobjects[0].name.namespace(), Some("cxx_qt"));
     }
+}
+
+/// Returns a vector of the names and types ([CppNamedType] of the parameters of this method, used in cpp generation step
+pub fn get_cpp_params(method: &ForeignItemFn, type_names: &TypeNames) -> Result<Vec<CppNamedType>> {
+    method
+        .sig
+        .inputs
+        .iter()
+        .map(|input| {
+            // Match parameters to extract their idents
+            if let FnArg::Typed(PatType { pat, ty, .. }) = input {
+                let ident = if let Pat::Ident(PatIdent { ident, .. }) = &**pat {
+                    ident
+                } else {
+                    return Err(Error::new(input.span(), "Unknown pattern for type"));
+                };
+
+                // If the name of the argument is self then ignore,
+                // as this is likely the self: Pin<T>
+                if ident == "self" {
+                    Ok(None)
+                } else {
+                    Ok(Some(CppNamedType {
+                        ident: ident.to_string(),
+                        ty: syn_type_to_cpp_type(ty, type_names)?,
+                    }))
+                }
+            } else {
+                Ok(None)
+            }
+        })
+        .filter_map(|result| result.map_or_else(|e| Some(Err(e)), |v| v.map(Ok)))
+        .collect()
 }
